% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc}
\alias{calc}
\title{calc: Calculate measure of association from numeric inputs}
\usage{
calc(calc, a, b = NA, c = NA, d = NA, n = NA, sumdiff = NA, conf_lvl = 0.95)
}
\arguments{
\item{calc}{The measure to be calculated. Choose from risk difference ("rd"), risk ratio ("rr"), incidence odds ratio ("ior"), incidence rate difference ("ird"), and incidence rate ratio ("irr"), or Mantel-Haenszel versions ("mhior", "mhrd", "mhrr").}

\item{a}{The number of exposed cases. Alternatively, a 2x2 dataframe or matrix, in format data.frame(c(a, c), c(b, d)) or matrix(c(a, c, b, d)), of counts.}

\item{b}{The number of exposed controls (or exposed person-time). Alternatively, if a is a dataframe or matrix, NA.}

\item{c}{The number of unexposed cases. Alternatively, if a is a dataframe or matrix, NA.}

\item{d}{The number of unexposed controls (or unexposed person-time). Alternatively, if a is a dataframe or matrix, NA.}

\item{n}{If a-d are actual counts of real persons (or person-time), NA. If a-d are sums of weights, n is the total number of actual persons sampled.}

\item{sumdiff}{If a-d are actual counts of real persons (or person-time), NA. If a-d are sums of weights, sumdiff is the sum of the squared differences between each observation's weight and the mean weight.}

\item{conf_lvl}{The preferred confidence level for hypothesis testing. Defaults to 0.95.}
}
\value{
A tibble.
}
\description{
Returns a tibble with a p-value, 95 percent CI, and estimated measure of association.
}
\details{
Recommended use: pipe to View("Title") for easiest viewing.

calc always returns the Wald method, including on Mantel-Haenszel measures.
}
