% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{.startup}
\alias{.startup}
\title{Internal function for startup checks}
\usage{
.startup(args, env)
}
\arguments{
\item{args}{A character vector of argument names to check}

\item{env}{An environment with the contents of arguments passed from the main function}

\item{...}{Additional arguments passed to the function}
}
\description{
Internal function for startup checks
}
\keyword{internal}
