% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcurve.R
\name{ezcurve}
\alias{ezcurve}
\title{ezcurve: Generate an epi curve}
\usage{
ezcurve(
  x,
  outcome_var,
  date_var,
  index_out = 1,
  start_date = NA,
  end_date = NA,
  title = NULL
)
}
\arguments{
\item{x}{A dataset.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{date_var}{A variable in x containing the date each outcome occurred.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{start_date}{The date on which to start the graph. Defaults to NA, which will automatically set the start date.}

\item{end_date}{The date on which to end the graph. Defaults to NA, which will automatically set the end date.}

\item{title}{The plot title. Defaults to NULL.}
}
\value{
A ggplot2 object.
}
\description{
Returns a ggplot2 ggplot object that, when displayed, is a bar chart of outcomes by their date of occurence.
}
\details{
You can follow the ezcurve call with further ggplot2 calls to customize the graph.
}
