% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezior.R
\name{ezior}
\alias{ezior}
\title{ezior: Odds ratio from binary exposure and outcome variables}
\usage{
ezior(
  x,
  exposure_var,
  outcome_var,
  index_exp = 1,
  ref_exp = 0,
  index_out = 1,
  ref_out = 0,
  conf_lvl = 0.95,
  print = TRUE
)
}
\arguments{
\item{x}{A dataset.}

\item{exposure_var}{A categorical exposure variable in x.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{index_exp}{The value of exposure_var to treat as the exposed group. Defaults to 1.}

\item{ref_exp}{The value of exposure_var to treat as the unexposed group. Defaults to 0.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}

\item{conf_lvl}{The preferred confidence level for hypothesis testing. Defaults to 0.95.}

\item{print}{Whether to print a counts table to the console. Defaults to TRUE.}
}
\value{
A tibble.
}
\description{
Returns a tibble with a p-value, 95 percent CI, and estimated incidence
odds ratio.
}
\details{
Recommended use: pipe to View("Title") for easiest viewing.
}
