% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezir.R
\name{ezir}
\alias{ezir}
\title{ezir: Table with case count, person-time totals, and incidence rates from
binary exposure and outcome variables and a person-time variable.}
\usage{
ezir(
  x,
  exposure_var,
  outcome_var,
  person_time,
  index_exp = 1,
  ref_exp = 0,
  index_out = 1,
  ref_out = 0
)
}
\arguments{
\item{x}{A dataset.}

\item{exposure_var}{A categorical exposure variable in x.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{person_time}{A variable in x reflecting the person-time contributed by each participant.}

\item{index_exp}{The value of exposure_var to treat as the exposed group. Defaults to 1.}

\item{ref_exp}{The value of exposure_var to treat as the unexposed group. Defaults to 0.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}
}
\value{
A tibble.
}
\description{
Returns a tibble with stratified counts, person-time totals, and rates by
exposure status.
}
