% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezprev.R
\name{ezprev}
\alias{ezprev}
\title{ezprev: Prevalence table from a binary outcome variable}
\usage{
ezprev(x, outcome_var, index_out = 1, ref_out = 0)
}
\arguments{
\item{x}{A dataset.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}
}
\value{
A tibble.
}
\description{
Returns a tibble with stratified counts by outcome, along with sums and
prevalence.
}
