% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztbl.R
\name{eztbl}
\alias{eztbl}
\title{eztbl: 2x2 contingency table from binary exposure and outcome variables}
\usage{
eztbl(
  x,
  exposure_var,
  outcome_var,
  index_exp = 1,
  ref_exp = 0,
  index_out = 1,
  ref_out = 0,
  risk = TRUE
)
}
\arguments{
\item{x}{A dataset.}

\item{exposure_var}{A categorical exposure variable in x.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{index_exp}{The value of exposure_var to treat as the exposed group. Defaults to 1.}

\item{ref_exp}{The value of exposure_var to treat as the unexposed group. Defaults to 0.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}

\item{risk}{If TRUE, calculates risks. If FALSE, calculates odds.}
}
\value{
A tibble.
}
\description{
Returns a tibble with stratified counts by exposure and outcome, along with
sums and risks or odds by exposure status.
}
