% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix}
\alias{matrix}
\title{matrix: Build counts frame to pass to calc}
\usage{
matrix(x.df)
}
\arguments{
\item{x.df}{A table generated by .standardize.}
}
\description{
Not exported. Takes standardized table and generates counts table. This
function used to generate a matrix, but doesn't anymore; the name was
maintained for consistency.
}
\keyword{internal}
