% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrr.R
\name{mhrr}
\alias{mhrr}
\title{mhrr: Mantel-Haenszel pooled risk ratio from binary exposure and outcome variables}
\usage{
mhrr(
  x,
  exposure_var,
  outcome_var,
  strat_var,
  index_exp = 1,
  ref_exp = 0,
  index_out = 1,
  ref_out = 0,
  conf_lvl = 0.95
)
}
\arguments{
\item{x}{A dataset.}

\item{exposure_var}{A categorical exposure variable in x.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{strat_var}{A categorical stratification variable in x to pool across.}

\item{index_exp}{The value of exposure_var to treat as the exposed group. Defaults to 1.}

\item{ref_exp}{The value of exposure_var to treat as the unexposed group. Defaults to 0.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}

\item{conf_lvl}{The preferred confidence level for hypothesis testing. Defaults to 0.95.}
}
\value{
A tibble.
}
\description{
Returns a tibble with a 95 percent CI and estimated risk ratio
(relative risk).
}
\details{
Recommended use: pipe to View("Title") for easiest viewing.
}
