% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreprev.R
\name{moreprev}
\alias{moreprev}
\title{moreprev: Prevalence table from a categorical outcome variable}
\usage{
moreprev(x, outcome_var)
}
\arguments{
\item{x}{A dataset.}

\item{outcome_var}{A categorical outcome variable in x.}
}
\value{
A tibble.
}
\description{
Returns a tibble with stratified counts by outcome, along with sums and
prevalence.
}
