% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morerr.R
\name{morerr}
\alias{morerr}
\title{morerr: Multiple risk ratios from categorical exposure and outcome variables}
\usage{
morerr(
  x,
  exposure_var,
  outcome_var,
  strat_var = NULL,
  ref_exp = 0,
  index_out = 1,
  ref_out = 0,
  conf_lvl = 0.95,
  mh = 0
)
}
\arguments{
\item{x}{A dataset.}

\item{exposure_var}{A categorical exposure variable in x.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{strat_var}{A categorical variable in x by which to stratify estimates.}

\item{ref_exp}{The value of exposure_var to treat as the unexposed group. Defaults to 0.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}

\item{conf_lvl}{The preferred confidence level for hypothesis testing. Defaults to 0.95.}

\item{mh}{Sets whether to use (Cochran-)Mantel-Haenszel pooling. 1 for yes, defaults to 0.}
}
\value{
A tibble.
}
\description{
Returns a tibble with stratified counts, risks, and ratios with associated CIs.
}
\details{
Recommended use: pipe to View("Title") for easiest viewing.
}
