% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretbl.R
\name{moretbl}
\alias{moretbl}
\title{moretbl: nx2 contingency table from categorical exposure and outcome variables}
\usage{
moretbl(x, exposure_var, outcome_var, ref_exp = 0, index_out = 1, ref_out = 0)
}
\arguments{
\item{x}{A dataset.}

\item{exposure_var}{A categorical exposure variable in x.}

\item{outcome_var}{A categorical outcome variable in x.}

\item{ref_exp}{The value of exposure_var to treat as the unexposed group. Defaults to 0.}

\item{index_out}{The value of outcome_var to treat as cases. Defaults to 1.}

\item{ref_out}{The value of outcome_var to treat as non-cases. Defaults to 0.}
}
\value{
A tibble.
}
\description{
Returns a tibble with stratified counts by exposure and outcome, along with
sums and risks by exposure status. Like eztbl, but allows for more exposure
categories (but still binary outcome).
}
\details{
Note: Before using moretbl, you must collapse all missing variables to NA.
All non-null values of exposure_var will be included in the analysis.
}
