% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_rows.R
\name{mutate_rows}
\alias{mutate_rows}
\title{mutate_rows: Create, modify, and delete rows}
\usage{
mutate_rows(x, ..., .numeric_data = FALSE)
}
\arguments{
\item{x}{A dataset.}

\item{...}{Name-value pairs and other arguments. See \code{\link[dplyr]{mutate}}.}

\item{.numeric_data}{Defaults to FALSE. When TRUE, data except the names column will be coerced to numeric.}
}
\value{
A tibble.
}
\description{
Like \code{mutate()} in \code{dplyr}, create new rows that are functions of existing rows. You can modify and delete rows by setting their value to null.
}
\details{
This function transposes the dataframe (using \code{ezt()}), passes its arguments to \code{mutate()}, then transposes it again. This means it works exactly as you would expect \code{mutate()} to work.

The first column of your dataset must be a unique identifier. Add one if necessary using \code{rowid_to_column()}.
}
