% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize: Standardize variables}
\usage{
standardize(args, list)
}
\arguments{
\item{args}{A string containing information about the arguments passed to the main function.}

\item{list}{A list of arguments passed to the main function.}
}
\description{
Not exported. Takes variables from original data and moves them to exposure and outcome variables.
}
\keyword{internal}
