% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{startup: Initial sanity checking}
\usage{
startup(args, list)
}
\arguments{
\item{args}{A string containing information about the arguments passed to the main function.}

\item{list}{A list of arguments passed to the main function.}
}
\description{
Not exported. Checks inputs for problems, halting execution as needed.
}
\keyword{internal}
