% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{table: Build counts table to pass to fmsb}
\usage{
table(x.df, index, risk, rate, print = FALSE)
}
\arguments{
\item{x.df}{A table generated by .standardize.}

\item{index}{If TRUE, exp is "exposed" or "unexposed". If FALSE, exp could have multiple indexes.}

\item{risk}{If TRUE, add a risk column. If FALSE, add an odds column. If NA, add prevalence or nothing.}

\item{rate}{If TRUE, generate a rate table. If FALSE, don't.}

\item{print}{Default is FALSE. If TRUE, prints the table to the console.}
}
\description{
Not exported. Takes standardized table and generates counts table for fmsb.
}
\keyword{internal}
