% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezt.R
\name{ezt}
\alias{ezt}
\title{ezt: Transpose a dataset}
\usage{
ezt(x, row_name = NULL)
}
\arguments{
\item{x}{A dataset.}

\item{row_name}{A string, which will be the name of the left-most column. Defaults to NULL, which will exclude the existing dataset's headers from the new dataset.}
}
\value{
A tibble.
}
\description{
Returns a tibble with horizontal and vertical axes reversed from the original dataset.
}
\details{
In some cases, the data you want to be your headers may be included as row names rather than a variable. If this is the case, pipe \code{as_tibble(x, rownames = "names")} into \code{ezt()}.
}
