% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix}
\alias{matrix}
\title{matrix: Build counts matrix to pass to fmsb}
\usage{
matrix(x.df, index, risk, rate)
}
\arguments{
\item{x.df}{A table generated by .standardize.}

\item{index}{If TRUE, exp is "exposed" or "unexposed". If FALSE, exp could have multiple indexes.}

\item{risk}{If TRUE, add a risk column. If FALSE, add an odds column. If NA, add prevalence or nothing.}

\item{rate}{If TRUE, generate a rate table. If FALSE, don't.}
}
\description{
Not exported. Takes standardized table and generates counts matrix for fmsb Mantel-Haenszel functions.
}
\keyword{internal}
